Concurrent threading methods tested and verified.
wake_up_sequence threading now runs as a daemon and is encapsulated by the WakeUpSequence class, which only exposes start and stopping functionality.

Basic logging implemented.

Sequence _should_ stop when an external source changes the lights. The xy property of the light is not stable (possibly the cause of the flickering I've sometimes noticed?). As a result, comparing the commanded xy to the actual xy of all the lights is unreliable. A small configurable dead band was added to account for this noise. Brightness and on/off state must be exact, however. Sequence interrupt was tested with Android app and Philips wall switch.